move train scss into ServiceAutocomplete scss
