--- src/components/TileLayout/AddMoreTile/Autocomplete/TrainAutoComplete/TrainAutoCompleteSelectLines/TrainAutoCompleteSelectLines.module.scss
@@ -1,3 +1,0 @@
-.lineName {
-  text-transform: capitalize;
-}
